requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
beautifulsoup4
arxiv
pymupdf
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain-huggingface
faiss-cpu
chromadb
langchain_chroma
beautifulsoup4
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette